requirements.txt
click
Deprecated
Flask
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
redis
six
Werkzeug
wrapt
zipp
mysql-connector-python
